chatbot-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
PyPDF2==3.0.1
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai>=0.5.0
